pyproject.toml
[tool.poetry]
name = "rasb_pi_timelapse"
version = "0.0.1"
description = ""
authors = ["Parametricall <22248105+Parametricall@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Pillow = "^7.2.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
